+++ test-utils/Cargo.toml
@@ -19,6 +19,7 @@ async-trait = "0.1"
 futures = { package = "futures-util", version = "0.3" }
 hex = "0.4"
 log = { package = "tracing", version = "0.1", features = ["log"] }
+num-traits = "0.2"
 rand = "0.8"
 rand_chacha = "0.3.1"
 parking_lot = "0.12"
