+++ test-utils/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod accounts_revert;
 pub mod block_production;
 pub mod blockchain;
 pub mod blockchain_with_rng;
